Update README.md with project details and usage instructions
The README.md file of the WebSocketDemo project has been expanded to provide a comprehensive project introduction and detailed instructions on how to use the project. This update includes steps for running the project and interacting with the WebSocket api using Apifox or an online WebSocket interface testing tool. Explanation of the usage of WebSocket interfaces for broadcasting chatroom join/leave notifications and for sending/receiving messages is now also included.